requirements.txt
gym~=0.17.3
matplotlib
pytorch~=1.6.0

numpy~=1.19.2
setuptools~=50.3.2
h5py~=2.10.0
pettingzoo~=1.3.5
stable-baselines3~=1.0